+++ src/app/config/api.config.ts
@@ -1,7 +1,11 @@
 export const API_CONFIG = {
   baseUrl: 'http://localhost:3000',
-  authToken: 'token-fake-aula-13-catalogo-disciplinas',
   endpoints: {
     campanhas: '/campanhas',
+    usuarios: '/usuarios',
+  },
+  storage: {
+    tokenKey: 'catalogo-disciplinas-token',
+    usuarioKey: 'catalogo-disciplinas-usuario',
   },
 };